requirements.txt
APScheduler==3.10.4
DateTime==5.5
mysql-connector-python==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.5
requests==2.32.3
websocket-client==1.8.0
websockets==12.0
pytz~=2024.1
